304: Move the CAN list React stuff into a sub-folder
