+++ frontend/src/index.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import App from "./App";
-import CanList from "./pages/cans/CanList";
+import CanList from "./pages/cans/list/CanList";
 import reportWebVitals from "./reportWebVitals";
 
 import { Provider } from "react-redux";
@@ -17,8 +17,9 @@ root.render(
             <BrowserRouter>
                 <Routes>
                     <Route path="/" element={<App />} />
-                    <Route path="/cans" element={<CanList />} />
-                    <Route path="/cans/:id" element={<App />} />
+                    <Route path="/cans" element={<CanList />}>
+                        <Route path="/cans/:id" element={<App />} />
+                    </Route>
                 </Routes>
             </BrowserRouter>
         </Provider>
