+++ frontend/src/pages/cans/list/CanList.jsx
