+++ frontend/src/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import canListSlice from "./pages/cans/canListSlice";
+import canListSlice from "./pages/cans/list/canListSlice";
 
 export default configureStore({
     reducer: {
